% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_data.R
\name{aggregate_data}
\alias{aggregate_data}
\title{aggregate_data}
\usage{
aggregate_data(df)
}
\description{
This function joins dataframe to the results identified in teh data aggregation process. It will subsittiute
IRResultNWQSunit for the mean IRResultNWQSunit of the group. It will keep one result, and discard the rest.
The source data, 'aggregate_data' needs to be updated every cycle.
}
\details{
@param df
 @export
}
