% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/air_station_lookup.R
\name{air_station_lookup}
\alias{air_station_lookup}
\title{air_station_lookup}
\usage{
air_station_lookup(
  df,
  monloc_col = "MLocID",
  Lat_col = "Lat_DD",
  Long_col = "Long_DD"
)
}
\arguments{
\item{df}{Dataframe of data or monitoring locations that need to be matched to air temp stations}

\item{monloc_col}{Name of column in df that contains monitoring location name data. Default is "MLOCID"}

\item{Lat_col}{Name of column in df that contains Latitude information. Default is "Lat_DD"}

\item{Long_col}{Name of column in df that contains Longitude information. Default is "Long_DD"}
}
\description{
This function matches monitoring stations to their nearest NOAA air temperature station. This tool references data
created by shp files generated from NOAA datasets that represent stations with enugh data to be usable for the air
temperature exclusion.
}
