% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/air_temp_exclusion.R
\name{air_temp_exclusion}
\alias{air_temp_exclusion}
\title{air_temp_exclusion}
\usage{
air_temp_exclusion(
  df,
  date_col = "Date",
  monloc_col = "MLocID",
  lat_col = "Lat_DD",
  long_col = "Long_DD"
)
}
\arguments{
\item{df}{dataframe to join air temperature data to. Intended to be an export from AWQMS or the IR database}

\item{date_col}{single element string identifying the name of the date column.}

\item{lat_col}{single element string identifying the name of the latitude column.}

\item{long_col}{single element string identifying the name of the longitude column.}

\item{mloc_col}{single element string identifying the name of the monitoring location column.}
}
\value{
Original dataframe with NOAA air temperature station ID, Name, and air temperature data appended
}
\description{
This function performs the air temperature exclusion check as identified in the Water Temperature standard.
A dataframe of data from AWQMS or the IR database is passed through the air_station_lookup function and then joined
with the OR_air_temp dataset which contains daily max air temp values (Air_Temp_daily_max) and calculated
90th percentile of annual seven-day average maximum air temperatures calculated using 10 years of air temperature
data (air_temp_exclusion_value). above_exclusion_1d identifies if that day's maximum air temperature was above
the air temp exclusion value. above_exclusion_7d identifies if any of the 7 previous days (inclusive of current day)
were above the air temperature exclusion value.
}
