% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Al_default_DOC.R
\name{Al_default_DOC}
\alias{Al_default_DOC}
\title{Aluminum default DOC Lookup}
\usage{
Al_default_DOC(lat, long)
}
\arguments{
\item{lat}{Latitude of station}

\item{long}{Longitude of station}
}
\value{
default DOC value. If -9999 is returned, station is outside of georeferenced area.
}
\description{
This function looks up default DOC values used in calculating Al criteria. The function sends lat/long values via
a query to a Oregon DEQ map server to determine deafult DOC values. DOC values are based on georegion as described in
##insert link to interpretation doc when finalized##. If -9999 is returned, station is outside of georeferenced area.
}
\examples{
\dontrun{
Al_default_DOC(lat = 45.80369 , long = -123.9551)

aluminum_data_mlocs <- aluminum_data \%>\%
 select(MLocID, Lat_DD, Long_DD) \%>\%
 distinct() \%>\%
 rowwise() \%>\%
 mutate(def_DOC = Al_default_DOC(Lat_DD, Long_DD))
}
}
