% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_delist.R
\name{assess_delist}
\alias{assess_delist}
\title{Assess_delist_ws}
\usage{
assess_delist(df, type = NULL)
}
\arguments{
\item{df}{dataframe of IR categories just after joining previous assessments}
}
\description{
This script will assign category 2 to GNIS's that meet a delisting condtion. If Delist_eligability == 1 and the
prev_GNIS_category %in% c('5',  '4A'), final_GNIS_cat will = 2. If  Delist_eligability != 1 and the
prev_GNIS_category %in% c('5',  '4A'), the final_GNIS_cat will equal the preivous category.
}
