% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_delisting.R
\name{binomial_delisting}
\alias{binomial_delisting}
\title{binomial_delisting}
\usage{
binomial_delisting(n, type = c("Toxics", "Conventionals"))
}
\arguments{
\item{n}{number of results.}

\item{type}{"Toxics" or "Conventionals" for parameter type.}
}
\value{
Value to delist if number of excursions is below.
}
\description{
This function returns the number of excursions required to delist. Using type = "Toxics" represents 90%
confidence that <= 5% samples exceed criterion value. type = "Conventionals" represnts 90% confidnce that <= 10%
samples exceed criterion value.
}
\examples{
binomial_delisting(30, type = "Toxics")
}
