% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_excursions.R
\name{binomial_excursions}
\alias{binomial_excursions}
\title{binomial_excursions}
\usage{
binomial_excursions(n, type = c("Toxics", "Conventionals"))
}
\arguments{
\item{n}{number of results.}

\item{type}{"Toxics" or "Conventionals" for parameter type.}
}
\value{
Value to list if number of excursions exceed.
}
\description{
This function returns the number of excursions required to list as impaired. Using type = "Toxics" represents 90%
confidence that <= 5% samples exceed criterion value. type = "Conventionals" represnts 90% confidnce that <= 10%
samples exceed criterion value.
}
\examples{
binomial_excursions(30, type = "Toxics")
}
