% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censor_data.R
\name{censor_data}
\alias{censor_data}
\title{Censor Data}
\usage{
censor_data(
  df,
  result_col = IRResultNWQSunit,
  qualifier_col = Result_Operator,
  criteria_col = crit
)
}
\arguments{
\item{df}{dataframe to modify.}

\item{result_col}{column in dataframe with the result to be compared against. Defaults to IRResultNWQSunit}

\item{qualifier_col}{column in dataframe for the qualifier. Defaults to Result_Operator}

\item{criteria_col}{column in dataframe with criteria value. Defaults to crit}
}
\description{
This function deals with censored data as described in the white paper "Use of Censored Data" from 5/11/2018
When the QL > criteria value, 1/2 the value of the criteria is substituted for any sample reported as censored
When the QL < Criteria value, 1/2 the QL value will be substituted
Samples greater than the Max QL, use QL value
Input is a dataframe
Adds a column named Result_cen for the value to use
}
