% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_split_AWQMS.R
\name{data_split_AWQMS}
\alias{data_split_AWQMS}
\title{Data Split function}
\usage{
data_split_AWQMS(df, split_on, size, filepath)
}
\arguments{
\item{df}{data frame to be split}

\item{split_on}{name of the column in dataframe to be split on. Datwithin this oclumn remains intact.}

\item{size}{row length to trigger splitting. Files will be longer than this length}

\item{filepath}{string for pathway to save file. Must end with /}
}
\description{
This function splits a data frame (df) into files of size n and saves csv to filepath
}
\examples{
data_split_AWQMS(NWIS_data, split_on = "SiteID", size = 100000, filepath = save_location)

}
