% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_AU_info.R
\name{join_AU_info}
\alias{join_AU_info}
\title{join_AUs}
\usage{
join_AU_info(df, AU_ID_col = "AU_ID")
}
\arguments{
\item{df}{dataframe to join}

\item{AU_ID_col}{Field that identifies AU_ID}
}
\description{
This function will join a dataframe by HUC12 number to HUC number and names for HUC 4 - 12. This function is used
to transfrom HUC12 in the geodatabase layer to additional HUC information to be used in tool sorting. The huc info
is stored in the HUC_info dataset
}
