% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_hucs.R
\name{join_hucs}
\alias{join_hucs}
\title{join_hucs}
\usage{
join_hucs(df, hu12_col = "HUC12")
}
\arguments{
\item{df}{dataframe to join}

\item{hu12_col}{Field that the original HUC12 number is identified}
}
\description{
This function will join a dataframe by HUC12 number to HUC number and names for HUC 4 - 12. This function is used
to transfrom HUC12 in the geodatabase layer to additional HUC information to be used in tool sorting. The huc info
is stored in the HUC_info dataset
}
