% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_prev_assessments.R
\name{join_prev_assessments}
\alias{join_prev_assessments}
\title{join_prev_assessments}
\usage{
join_prev_assessments(df, AU_type)
}
\arguments{
\item{df}{Dataframe of new assessment conclusions}

\item{AU_type}{What sort of AU type is the join opertaing on "WS" or Other"}
}
\description{
This function joins in previous assessment conclusions to current assessment process.
If AY_type == 'WS', the data will be joined to previous impairements bu AU_GNIS as well as overall AU categorization.
IF AU_type is 'other', the data will be jopined to overall AU assessment conclusion.
}
