% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_TMDL.R
\name{join_TMDL}
\alias{join_TMDL}
\title{join_TMDL}
\usage{
join_TMDL(df, type)
}
\arguments{
\item{df}{dataframe to join. Shoud;d either be final AU decsion table or final GNIS decision table, after
all delist etc completed}

\item{type}{should be 'AU' or 'GNIS' depending on what you ware matching}
}
\value{
Returns same datafrane as df, but with action ID and TMDL parameters added
}
\description{
Join assessments to TMDL actions. used to determine categery 4A and to display action IDs in display
}
