% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_datetimes.R
\name{normalize_datetime}
\alias{normalize_datetime}
\title{Normalize Datetimes}
\usage{
normalize_datetime(
  df,
  datecol = "SampleStartDate",
  timecol = "SampleStartTime",
  tzcol = "SampleStartTZ",
  target_tz = "America/Los_Angeles",
  norm_time_colname = "datetime",
  vector_only = FALSE
)
}
\arguments{
\item{df}{dataframe to modify.}

\item{datecol}{Name of the column that contains date values.}

\item{timecol}{Name of the column that contains time values.}

\item{tzcol}{Name of column that contains timezone values.}

\item{target_tz}{Timezone data should be normalized to.}

\item{norm_time_colname}{Name of column new data should be written to.}

\item{vector_only}{If true, return only a vector of normalized datetimes. IF false, return dataframe with added datetime}
}
\description{
This function takes in datetime data that is collected under different timezones, and normalizes them to the same
time zone. This function is needed because R does not allow multiple timezones in a vector.
}
