% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NWIS_cont_data_pull.R
\name{NWIS_cont_data_pull}
\alias{NWIS_cont_data_pull}
\title{NWIS_cont_data_pull}
\usage{
NWIS_cont_data_pull(start.date, end.date, save_location, stateCD = "or")
}
\arguments{
\item{start.date}{Start date of data pull}

\item{end.date}{End date of datapull}

\item{save_location}{Folder to save output files to}

\item{stateCD}{Statecode for datapull. Defaults to 'or'}
}
\description{
This function queries continuous data from the NWIS database and produces AWQMS import files. Temperature, pH, and DO
are queried and IR summary statistics are calculated. THe function outputs three csv files: NWIS Sumamry Stats,
NWIS continuous pH, and NWIS monitoring location information. The function has no outputs, the CSV files serve as
outputs.
}
