% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odeqIRtools-package.R
\docType{package}
\name{odeqIRtools-package}
\alias{odeqIRtools}
\alias{odeqIRtools-package}
\title{odeqIRtools: Tools used in Oregon DEQ's Integrated Report Assessment}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/TravisPritchardODEQ/odeqIRtools}
  \item Report bugs at \url{https://github.com/TravisPritchardODEQ/odeqIRtools/issues}
}

}
\keyword{internal}
