% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rollup_WS_AU.R
\name{rollup_WS_AU}
\alias{rollup_WS_AU}
\title{rollup_WS_AU}
\usage{
rollup_WS_AU(df, char_name_field)
}
\arguments{
\item{df}{input dataframe of GNIS assessments}

\item{char_name_field}{field that contains the characteristic name.No quotes.}
}
\description{
This function rolls up GNIS assessments to the AU scale.
}
