% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strd_lyr_display.R
\name{strd_lyr_display}
\alias{strd_lyr_display}
\title{Creates the csv file needed for displaying the standards info on the webmap.}
\usage{
strd_lyr_display(GeoRef_Standards_path, relation_table, write_path)
}
\arguments{
\item{GeoRef_Standards_path}{The Georeferenced standards layer. Use the GDB file in the GIS library}

\item{relation_table}{The excel doc with the relationship tables for the standards layer. Should be on the sharepoint site}

\item{write_path}{The path to save the outputs}
}
\value{
Returns nothing. Saves 2 Rdata files and the strd layer display csv to write_path
}
\description{
This function pulls the standards information from the georeferenced standards layer and joins
the codes to the relaionship table. This is the data source for when you click on the standards
layer and you get the popup with all the standards in one place. For 2018 and 2022 this is passed to
Dave Pray to integrate into the standards layer.
}
