% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique_AU.R
\name{unique_AU}
\alias{unique_AU}
\title{unique_AU}
\usage{
unique_AU(AU_ID)
}
\arguments{
\item{AU_ID}{AU_ID string to extract unique code portion from}
}
\value{
THe unique portion of an AU_ID as string
}
\description{
This function extracts the unique element from an AU_ID. It will extract the last element defines by the '_' symbol.
}
\examples{
unique_AU('OR_WS_170501100706_05_102965')
}
