% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WS_AU_prev_list.R
\name{WS_AU_prev_list}
\alias{WS_AU_prev_list}
\title{WS_AU_prev_list}
\usage{
WS_AU_prev_list(df)
}
\arguments{
\item{df}{input dataframe. Usually the output of rollup_WS_AU()}
}
\description{
This function adds previous cycle AU assessments to GNIS assessments taht have been rolled up to the AU scale.
}
